backend/requirements.txt
fastapi
uvicorn
SmartApi
python-dotenv
pyotp
requests
websocket-client
